compose: Add types to createHigherOrderComponent and ifCondition
