+++ packages/compose/tsconfig.json
@@ -1,0 +1,11 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"declarationDir": "build-types"
+	},
+	"include": [
+		"src/higher-order/if-condition/**/*",
+		"src/utils/**/*"
+	]
+}
